app/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
defusedxml==0.7.1
Django==5.2.5
django-debug-toolbar==6.0.0
django-extensions==4.1
django-recaptcha==4.1.0
django-taggit==6.1.0
djangorestframework==3.16.1
environs==14.3.0
Faker==37.8.0
gevent==25.9.1
git-filter-repo==2.47.0
greenlet==3.2.4
idna==3.10
kombu==5.5.4
marshmallow==4.0.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
PyJWT==2.10.1
pymemcache==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
redis==5.2.1
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
zope.event==6.0
zope.interface==8.0.1
gunicorn==21.2.0